added onSnapShot firebase listener to the posts
