+++ src/screens/PostScreen.js
@@ -27,7 +27,7 @@ const PostScreen = ({ navigation, route }) => {
         mediaTypes: ImagePicker.MediaTypeOptions.Images,
         allowsEditing: true,
         aspect: [4, 3],
-        quality: 1,
+        quality: 0,
       });
 
       if (!result.canceled) {
