+++ src/screens/ProfileScreen.js
@@ -93,7 +93,7 @@ const ProfileScreen = ({ navigation }) => {
         mediaTypes: ImagePicker.MediaTypeOptions.Images,
         allowsEditing: true,
         aspect: [1, 1],
-        quality: 1,
+        quality: 0,
       });
 
       if (!result.canceled) {
